% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aocp.R
\name{aocp}
\alias{aocp}
\title{Formula for calculating the Average Oil Content Percentage (AOCP)}
\usage{
aocp(ocs1, ocs2)
}
\arguments{
\item{ocs1}{Oil content sample1 percentage}

\item{ocs2}{Oil content sample1 percentage}
}
\value{
Return the average oil content percentage 'aocp'
}
\description{
Formula for calculating the Average Oil Content Percentage (AOCP)
}
\details{
Formula for calculating the average oil content percentage
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{atw}},
  \code{\link{avdm}}, \code{\link{dm1}}, \code{\link{dm2}},
  \code{\link{ocs1}}, \code{\link{ocs2}}, \code{\link{sg}}
}
\author{
Omar Benites
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}
