% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmw.R
\name{atmw}
\alias{atmw}
\title{Formula for calculating the Average Marketable Tuber Weight (ATMW)}
\usage{
atmw(mtwp, nmtp)
}
\arguments{
\item{mtwp}{marketable tuber weight/plot}

\item{nmtp}{number marketable tubers/plot}
}
\value{
Return the average marketable tuber weight
}
\description{
Formula for calculating the Average Marketable Tuber Weight (ATMW)
}
\details{
Formula for calculating the average marketable tuber weight
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest,plant: \code{\link{mtya}},
  \code{\link{nmtpl}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

