% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmw_1hd.R
\name{atmw_1hd}
\alias{atmw_1hd}
\title{Function for calculating Average marketable tuber weight - First harvest date  (ATMW_1HD)}
\usage{
atmw_1hd(mtwp_1hd, nmtp_1hd)
}
\arguments{
\item{mtwp_1hd}{Number marketable tubers/plot  (First harvest date)}

\item{nmtp_1hd}{Marketable tuber weight/plot (First harvest date)}
}
\value{
atmw_1hd Returns the average marketable tuber weight (First harvest date)
}
\description{
Function for calculating Average marketable tuber weight - First harvest date  (ATMW_1HD)
}
\details{
This function returns average marketable tuber weight (First harvest date)
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_2hd}},
  \code{\link{atmw_3hd}}, \code{\link{atnomw_1hd}},
  \code{\link{atnomw_2hd}}, \code{\link{atnomw_3hd}},
  \code{\link{atnomw}}, \code{\link{av_sg}},
  \code{\link{sg_2hd}}, \code{\link{sg_3hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

