% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmw_2hd.R
\name{atmw_2hd}
\alias{atmw_2hd}
\title{Function for calculating Average marketable tuber weight-Second harvest date (ATMW_2HD)}
\usage{
atmw_2hd(mtwp_2hd, nmtp_2hd)
}
\arguments{
\item{mtwp_2hd}{Marketable tuber weight/plot (Second harvest date)}

\item{nmtp_2hd}{Number marketable tubers/plot  (Second harvest date)}
}
\value{
atmw_2hd Return the Average marketable tuber weight-Second harvest date
}
\description{
Function for calculating Average marketable tuber weight-Second harvest date (ATMW_2HD)
}
\details{
This function returns average marketable tuber weight-Second harvest date
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_3hd}}, \code{\link{atnomw_1hd}},
  \code{\link{atnomw_2hd}}, \code{\link{atnomw_3hd}},
  \code{\link{atnomw}}, \code{\link{av_sg}},
  \code{\link{sg_2hd}}, \code{\link{sg_3hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}
\author{
Omar Benites
}
