% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/atmw_3hd.R
\name{atmw_3hd}
\alias{atmw_3hd}
\title{Function for calculating Average marketable tuber weight - First harvest date  (ATMW_3HD)}
\usage{
atmw_3hd(mtwp_3hd, nmtp_3hd)
}
\arguments{
\item{mtwp_3hd}{Number marketable tubers/plot  (First harvest date)}

\item{nmtp_3hd}{Marketable tuber weight/plot (First harvest date)}
}
\value{
atmw_3hd Returns the average marketable tuber weight (First harvest date)
}
\description{
Function for calculating Average marketable tuber weight - First harvest date  (ATMW_3HD)
}
\details{
This function returns average marketable tuber weight (First harvest date)
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2034
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}};
  \code{\link{atmw_2hd}}; \code{\link{atnomw_1hd}};
  \code{\link{atnomw_2hd}}; \code{\link{atnomw_3hd}};
  \code{\link{sg_1hd}}; \code{\link{sg_2hd}};
  \code{\link{sg_3hd}}; \code{\link{sgs1_1hd}};
  \code{\link{sgs1_2hd}}; \code{\link{sgs1_3hd}};
  \code{\link{sgs2_1hd}}; \code{\link{sgs2_2hd}};
  \code{\link{sgs2_3hd}}
}

