% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atnomw.R
\name{atnomw}
\alias{atnomw}
\title{Function for calculating the Average Non-Marketable Tuber Weight}
\usage{
atnomw(nomtwp, nnomtp)
}
\arguments{
\item{nomtwp}{Non-marketable tuber weight/plot at harvest date}

\item{nnomtp}{Number of non-marketable tubers/plot at harvest date}
}
\value{
atnomw Return the average non-marketable tuber weight at harvest date
}
\description{
Function for calculating the Average Non-Marketable Tuber Weight
}
\details{
This function returns the average non-marketable tuber weight.
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 1014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{atnomw_3hd}}, \code{\link{av_sg}},
  \code{\link{sg_2hd}}, \code{\link{sg_3hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

