% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atnomw_1hd.R
\name{atnomw_1hd}
\alias{atnomw_1hd}
\title{Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_1HD)}
\usage{
atnomw_1hd(nomtwp_1hd, nnomtp_1hd)
}
\arguments{
\item{nomtwp_1hd}{Non-marketable tuber weight/plot  (Second harvest date)}

\item{nnomtp_1hd}{Number of non-marketable tubers/plot  (Second harvest date)}
}
\value{
atnomw_1hd Return the average non-marketable tuber weight (Second harvest date)
}
\description{
Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_1HD)
}
\details{
This function returns the average non-marketable tuber weight-Second harvest date
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 1014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_2hd}}, \code{\link{atnomw_3hd}},
  \code{\link{sg_1hd}}, \code{\link{sg_2hd}},
  \code{\link{sg_3hd}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

