% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atnomw_2hd.R
\name{atnomw_2hd}
\alias{atnomw_2hd}
\title{Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_2HD)}
\usage{
atnomw_2hd(nomtwp_2hd, nnomtp_2hd)
}
\arguments{
\item{nomtwp_2hd}{Non-marketable tuber weight/plot  (Second harvest date)}

\item{nnomtp_2hd}{Number of non-marketable tubers/plot  (Second harvest date)}
}
\value{
atnomw_2hd Return the average non-marketable tuber weight (Second harvest date)
}
\description{
Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_2HD)
}
\details{
This function returns the average non-marketable tuber weight-Second harvest date
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_3hd}},
  \code{\link{atnomw}}, \code{\link{av_sg}},
  \code{\link{sg_2hd}}, \code{\link{sg_3hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}
\author{
Omar Benites
}
