% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atnomw_3hd.R
\name{atnomw_3hd}
\alias{atnomw_3hd}
\title{Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_3HD)}
\usage{
atnomw_3hd(nomtwp_3hd, nnomtp_3hd)
}
\arguments{
\item{nomtwp_3hd}{Non-marketable tuber weight/plot  (Second harvest date)}

\item{nnomtp_3hd}{Number of non-marketable tubers/plot  (Second harvest date)}
}
\value{
atnomw_3hd Return the average non-marketable tuber weight (Second harvest date)
}
\description{
Function for calculating the Average non-marketable tuber weight-Second harvest date (ATNoMW_3HD)
}
\details{
This function returns the average non-marketable tuber weight-Second harvest date
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 3034
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{sg_1hd}}, \code{\link{sg_2hd}},
  \code{\link{sg_3hd}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

