% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atw.R
\name{atw}
\alias{atw}
\title{Formula for calculating the Average Tuber Weight (ATW)}
\usage{
atw(ttwp, tntp)
}
\arguments{
\item{ttwp}{Total tuber weight/plot}

\item{tntp}{Total number of tubers/plot}
}
\value{
Return the average tuber weight 'atw'
}
\description{
Formula for calculating the Average Tuber Weight (ATW)
}
\details{
Formula for calculating average tuber weight
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{aocp}},
  \code{\link{avdm}}, \code{\link{dm1}}, \code{\link{dm2}},
  \code{\link{ocs1}}, \code{\link{ocs2}}, \code{\link{sg}}
}
\author{
Omar Benites
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}
