% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_sg.R
\name{av_sg}
\alias{av_sg}
\title{Function for calculating Average Specific Gravity}
\usage{
av_sg(...)
}
\arguments{
\item{sgs_1hd}{Specific gravity sample 1 (First harvest date)}

\item{sgs2_1hd}{Specific gravity sample 2 (First harvest date)}
}
\value{
sg_1hd Return the specific gravity (First harvest date)
}
\description{
Function for calculating Average Specific Gravity
}
\details{
This function returns the specific gravity (First harvest date)
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{atnomw_3hd}}, \code{\link{atnomw}},
  \code{\link{sg_2hd}}, \code{\link{sg_3hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

