% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/avdm.R
\name{avdm}
\alias{avdm}
\title{Formula for calculating the Average Dry Matter (AVDM)}
\usage{
avdm(dm1, dm2)
}
\arguments{
\item{dm1}{Dry matter content sample 1}

\item{dm2}{Dry matter content sample 2}
}
\value{
Return the average dry matter
}
\description{
Formula for calculating the Average Dry Matter (AVDM)
}
\details{
Formula for calculating the average dry matter
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{aocp}};
  \code{\link{atw}}; \code{\link{dm1}}; \code{\link{dm2}};
  \code{\link{ocs1}}; \code{\link{ocs2}}; \code{\link{sg}}
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

