% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/avlglatsp.R
\name{avlglatsp}
\alias{avlglatsp}
\title{Function for calculating the Average Length of lateral sprout (AVLGLATSP)}
\usage{
avlglatsp(lglatsp1, lglatsp2, lglatsp3, lglatsp4, lglatsp5, lglatsp6)
}
\arguments{
\item{lglatsp1}{Length of the lateral sprout Number 1}

\item{lglatsp2}{Length of the lateral sprout Number 2}

\item{lglatsp3}{Length of the lateral sprout Number 3}

\item{lglatsp4}{Length of the lateral sprout Number 4}

\item{lglatsp5}{Length of the lateral sprout Number 5}

\item{lglatsp6}{Length of the lateral sprout Number 6}
}
\value{
avlgatsp
}
\description{
This functions calculate the average length of lateral sprout
}
\author{
Omar Benites
}
\references{
Procedures for standard evaluation and data management of advanced potato clones: Assessment of Dormancy and Sprouting
Behavior of Elite and Advanced Clones.International Potato Center (CIP).
}
\seealso{
Other Sprout-dormancy, evaluation, potato: \code{\link{avthsp}};
  \code{\link{dorpd}}; \code{\link{pw_spt}};
  \code{\link{pw_uspt}}
}

