% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochemal_content.R
\name{biochem_cont}
\alias{biochem_cont}
\title{Calculate the biochemical content for micro and macro nutrients}
\usage{
biochem_cont(biochem, dm)
}
\arguments{
\item{biochem}{biochemical element}

\item{dm}{dry matter}
}
\value{
Return the biochemichal content
}
\description{
Calculate the biochemical content for micro and macro nutrients
}
\details{
calculation of the biochemical content
}
\references{
Assessing potato clone for quality and nutrition
}
\author{
Omar Benites
}
