% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{calculate_form_dormancy}
\alias{calculate_form_dormancy}
\title{Calculation of standard evaluation values in Dormancy Test in Potato Seed Tubers}
\usage{
calculate_form_dormancy(data)
}
\arguments{
\item{data}{A dataframe with the fieldbook data}
}
\description{
To determine the dormancy period and the initiation of sprouting of advanced
clones for its use as seed under storage conditions of diffuse light or rustic storage.
}

