% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{calculate_form_harvest}
\alias{calculate_form_harvest}
\title{Calculation of standard evaluation of yield variables under Participatory Varietal Selection Methodology}
\usage{
calculate_form_harvest(data, plot_size = NA, plant_den = NA)
}
\arguments{
\item{data}{A data frame with the fieldbook data}

\item{plot.size}{The plot size in square meters}

\item{plant.den}{Plant density}
}
\description{
During the harvest period, agricultors and researchers make direct observations or measurements by counting and weighing
yield traits in the field. This evaluation is carried out for each clone and plot (each repetition of the Mother trial / each Baby trial).
The amount of harvested plants and the number and weight of the tubers are recorded
}

