% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{calculate_form_postharvest}
\alias{calculate_form_postharvest}
\title{Percetages and Global scores by Selection Criteria at Post-Harvest Stage}
\usage{
calculate_form_postharvest(data)
}
\arguments{
\item{data}{A dataframe with the fieldbook data}
}
\description{
With the group of farmers, the storage facilities of each one of the families is visited in order to
select the best clones according to storage characteristics, taking into account the previously
identified criteria. If desirable, the clones / varieties in each of the 3 farmer stores can also be
ranked by different groups of evaluators
}

