% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{calculate_form_sclones}
\alias{calculate_form_sclones}
\title{Mother and Baby scores calculations to select best clones by Parcel.}
\usage{
calculate_form_sclones(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
With the group of farmers, all of the (previously) harvested fields are visited in order to select 
the best clones, taking into account the previously identified criteria. In the case of the Mother 
trial, the evaluation is performed for each repetition.
}

