% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{calculate_form_scriteria}
\alias{calculate_form_scriteria}
\title{Percetages and global scores by Selection Criteria}
\usage{
calculate_form_scriteria(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Gathering and Ranking of Criteria at Flowering, harvest and post-harvest stage. 
The group is gathered and the objectives of the trial and the evaluation are briefly explained.
The group is asked: What is it that you look for in a new variety of potato, taking into account
its foliage? In other words: When do we say that a variety is good, while evaluating only its
foliage?.
The greatest amount of possible answers are encouraged and a list is compiled of all the
criteria and reasons mentioned by the farmers.
For example:
(a) Resistance to late blight,
(b) Enough foliage to feed my cow,
(c) That the plant does not extend and touch the floor (size when erect),
(d) That it is resistant to moths,
(e) That the foliage is very green,
(f) That the plant shows vigor, etc. (according to what is indicated by the farmers).
Each of the mentioned criteria is written on a paper bag or on a cardboard tray. In other
words, if the farmers have identified 6 criteria, then we also have 6 bags or cardboard trays.
}

