% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Formula for calculating the Percent Marketable Roots - Commercial Index (CI)}
\usage{
ci(nocr, nonc)
}
\arguments{
\item{nocr}{Number of commercial roots}

\item{nonc}{Number of non commercial roots}
}
\value{
Return percent marketable roots - commercial index 'ci'
}
\description{
Formula for calculating the Percent Marketable Roots - Commercial Index (CI)
}
\details{
Formula for calculating the percent marketable roots - commercial index
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{cytha}};
  \code{\link{dmfy}}; \code{\link{dmry}}; \code{\link{dm}};
  \code{\link{fytha}}; \code{\link{hi}};
  \code{\link{nrpp}}; \code{\link{rfr}};
  \code{\link{rytha}}; \code{\link{trw}}; \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

