% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cip_number.R
\name{cip_number_check}
\alias{cip_number_check}
\title{Function to detect if a cipnumer (the passport data) is well written following the CIP syntax}
\usage{
cip_number_check(cipnumber)
}
\arguments{
\item{cipnumber}{The cipnumber data.}
}
\value{
A list of values which are correct cipnumber or not.
}
\description{
Function to detect if a cipnumer (the passport data) is well written following the CIP syntax
}
\details{
A cipnumber is the formal name for every genotipe or variety which has been developed and CIP centers.
The syntax is (CIP)(1 until 6 numbers)(dot)(1 until max 3 digits)
}
\author{
Omar Benites
}
\keyword{cipnumber,}

