% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cipnumber_out_excel.R
\name{cipnumber_out}
\alias{cipnumber_out}
\title{Function to highlight CIPNUMBERS which are not well written (MSExcel)}
\usage{
cipnumber_out(file, cip_number, sheetname = "mysheet")
}
\arguments{
\item{file}{the direccion of the file}

\item{cip_number}{The cipnumber column name. (Mostly the genotipe column name)}

\item{sheetname}{The sheetname of the fieldbook}
}
\value{
The cipnumbers highlighted by color
}
\description{
Function to highlight CIPNUMBERS which are not well written (MSExcel)
}
\details{
This function allow detect cipnumber which are not correct, according to International Potato Center
nomeclatures.
}
\author{
Omar Benites
}
\keyword{cipnumer,}
\keyword{color,}
\keyword{correct,}
\keyword{highlight,}
\keyword{incorrect}

