% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm}
\alias{dm}
\title{Formula for calculating the Percentage of Storage Root Dry Matter Content (DM)}
\usage{
dm(dmd, dmf)
}
\arguments{
\item{dmd}{Dry weight of dmf samples}

\item{dmf}{Fresh weight of roots for dry matter assessment}
}
\value{
Return the percentage of storage root dry matter content 'dm'
}
\description{
Formula for calculating the Percentage of Storage Root Dry Matter Content (DM)
}
\details{
Formula for calculating the percentage of storage root dry matter content
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}},
  \code{\link{biom}}, \code{\link{ci}},
  \code{\link{cytha}}, \code{\link{dmfy}},
  \code{\link{dmry}}, \code{\link{fytha}},
  \code{\link{hi}}, \code{\link{nrpp}}, \code{\link{rfr}},
  \code{\link{rytha}}, \code{\link{trw}}, \code{\link{ypp}}
}
\author{
Omar Benites
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}
