% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dm1.R
\name{dm1}
\alias{dm1}
\title{Formula for calculating the Dry Matter Content Sample 1 (DM1)}
\usage{
dm1(dwts1, fwts1)
}
\arguments{
\item{dwts1}{Dry weight of tuber sample 1}

\item{fwts1}{Fresh weight of tuber sample 1}
}
\value{
Return the dry matter content sample 1
}
\description{
Formula for calculating the Dry Matter Content Sample 1 (DM1)
}
\details{
Formula for calculating the dry matter content sample 1
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{aocp}};
  \code{\link{atw}}; \code{\link{avdm}}; \code{\link{dm2}};
  \code{\link{ocs1}}; \code{\link{ocs2}}; \code{\link{sg}}
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

