% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dm2.R
\name{dm2}
\alias{dm2}
\title{Formula for calculating the Dry Matter Content Sample 2 (DM2)}
\usage{
dm2(dwts2, fwts2)
}
\arguments{
\item{dwts2}{Dry weight of tuber sample 2}

\item{fwts2}{Fresh weight of tuber sample 2}
}
\value{
Return the dry matter content sample 2
}
\description{
Formula for calculating the Dry Matter Content Sample 2 (DM2)
}
\details{
Formula for calculating the dry matter content sample 2
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{aocp}};
  \code{\link{atw}}; \code{\link{avdm}}; \code{\link{dm1}};
  \code{\link{ocs1}}; \code{\link{ocs2}}; \code{\link{sg}}
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

