% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dmfy.R
\name{dmfy}
\alias{dmfy}
\title{Formula for calculating the Dry matter foliage yield (DMFY)}
\usage{
dmfy(vw, pls, dmvd, dmvf)
}
\arguments{
\item{vw}{Vine weight}

\item{pls}{net plot size}

\item{dmvd}{Dry weight of DMVD samples}

\item{dmvf}{Fresh weight vines for dry matter assessment}
}
\value{
Return the root foliage ratio 'dmfy'
}
\description{
Formula for calculating the Dry matter foliage yield (DMFY)
}
\details{
Formula for calculating dry matter foliage yield
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{ci}};
  \code{\link{cytha}}; \code{\link{dmry}};
  \code{\link{dm}}; \code{\link{fytha}}; \code{\link{hi}};
  \code{\link{nrpp}}; \code{\link{rfr}};
  \code{\link{rytha}}; \code{\link{trw}}; \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

