% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dmry.R
\name{dmry}
\alias{dmry}
\title{Formula for calculating the Dry matter root  yield (DMRY)}
\usage{
dmry(crw, ncrw, pls, dmd, dmf)
}
\arguments{
\item{crw}{Commercial root weight}

\item{ncrw}{Non commercial root weight}

\item{pls}{Net plot size}

\item{dmd}{Dry weight of dmf samples}

\item{dmf}{Fresh weight of roots for dry matter assessment}
}
\value{
Return the Dry matter root yield 'dmry'
}
\description{
Formula for calculating the Dry matter root  yield (DMRY)
}
\details{
Formula for calculating dry matter root yield
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{ci}};
  \code{\link{cytha}}; \code{\link{dmfy}};
  \code{\link{dm}}; \code{\link{fytha}}; \code{\link{hi}};
  \code{\link{nrpp}}; \code{\link{rfr}};
  \code{\link{rytha}}; \code{\link{trw}}; \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

