% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dorpd.R
\name{dorpd}
\alias{dorpd}
\title{Function for calculating the Number Date of dehaulming to Date of sprouting (DORPD)}
\usage{
dorpd(datesp, ddehaulm)
}
\arguments{
\item{datesp}{Date of sprouting (format: yyyy-mm-dd)}

\item{ddehaulm}{Date of dehaulming (format: yyyy-mm-dd)}
}
\value{
Return the
}
\description{
Counted as days from haulm_cutting to sprouting of 80% of the tubers with at least one sprout longer than 2 mm.
Monitoring of sprout initiation and growth should be carried out at 10 day-intervals in order to record an accurate dormancy period
}
\details{
This function returns
}
\author{
Omar Benites
}
\references{
Procedures for standard evaluation and data management of advanced potato clones: Assessment of Dormancy and Sprouting
Behavior of Elite and Advanced Clones.International Potato Center (CIP).
}
\seealso{
Other Sprout-dormancy, evaluation, potato: \code{\link{avlglatsp}};
  \code{\link{avthsp}}; \code{\link{pw_spt}};
  \code{\link{pw_uspt}}
}

