% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drought_index}
\alias{drought_index}
\title{Drought Indexes based on means and summary sheet, both previously calculated by HIDAP}
\usage{
drought_index(fb, f_lvl1, f_lvl2)
}
\arguments{
\item{fb}{the fieldbook data}

\item{f_lvl1}{factor's first level}

\item{f_lvl2}{factor's second level}
}
\description{
Calcutation of drought indexes using two levels in treatments
if just few or neither elements are contained into the set.
}
\author{
Omar Benites
}

