% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DSI.mean.R
\name{DSI.means}
\alias{DSI.means}
\title{Function to calculate the Drought Suceptibility Index (DSI) using mean yield values}
\usage{
DSI.means(ylmeans, clabel = NA, slabel = NA)
}
\arguments{
\item{ylmeans}{A data frame with yield means. Included genotypes and Factor.}

\item{clabel}{The label of water control factor. Example: "irrigation","normal irrigation"}

\item{slabel}{The label of water stress factor. Example: "stress", "inttermitent}
}
\value{
The drought suceptible index using means
}
\description{
Function to calculate the Drought Suceptibility Index (DSI) using mean yield values
}
\details{
This function returns the drought suceptible index (DSI) using means.
}
\author{
Omar Benites
}
\references{
Comparison of yield based drought tolerance indices in improved varieties, genetic stocks and landraces
of potato (Solanum tuberosum L.)
}
\keyword{drought}
\keyword{index,}

