% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Formula for calculating the Harvest index (HI)}
\usage{
hi(crw, ncrw, vw)
}
\arguments{
\item{crw}{Commercial root weight}

\item{ncrw}{Non commercial root weight}

\item{vw}{Vine weight}
}
\value{
Return the harvest index 'hi'
}
\description{
Formula for calculating the Harvest index (HI)
}
\details{
Formula for calculating harvest index
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{ci}};
  \code{\link{cytha}}; \code{\link{dmfy}};
  \code{\link{dmry}}; \code{\link{dm}};
  \code{\link{fytha}}; \code{\link{nrpp}};
  \code{\link{rfr}}; \code{\link{rytha}};
  \code{\link{trw}}; \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

