% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INChC1.R
\name{INChC1}
\alias{INChC1}
\title{Function for calculating the Increase Chlorophyll Content 1 - First difference (INChC1)}
\usage{
INChC1(ChC2, ChC1)
}
\arguments{
\item{ChC2}{Chlorophyll Content sample 2}

\item{ChC1}{Chlorophyll Content sample 1}
}
\value{
Return the Increase Chlorophyll Content 1

Return the Increase Chlorophyll Content 2

Return the Increase Chlorophyll Content 3

Return the Increase Chlorophyll Content 4
}
\description{
Function for calculating the Increase Chlorophyll Content 1 - First difference (INChC1)
}
\details{
This function calculates the Increase Chlorophyll Content 1 (First difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC2}},
  \code{\link{INChC3}}, \code{\link{INChC4}},
  \code{\link{INRWC1}}, \code{\link{INRWC2}},
  \code{\link{RWC_EV1}}, \code{\link{RWC_EV2}},
  \code{\link{RWC_EV3}}, \code{\link{inplahe1}},
  \code{\link{inplahe2}}, \code{\link{inplahe3}},
  \code{\link{insd1}}, \code{\link{insd2}},
  \code{\link{insnpp1}}, \code{\link{insnpp2}}
}

