% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/INChC2.R
\name{INChC2}
\alias{INChC2}
\title{Function for calculating the Increase Chlorophyll Content 2 - Second difference (INChC2)}
\usage{
INChC2(ChC3, ChC2)
}
\arguments{
\item{ChC3}{Chlorophyll Content sample 3}

\item{ChC2}{Chlorophyll Content sample 2}
}
\value{
Return the Increase Chlorophyll Content 2
}
\description{
Function for calculating the Increase Chlorophyll Content 2 - Second difference (INChC2)
}
\details{
This function calculates the Increase Chlorophyll Content 2 (Second difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC3}}; \code{\link{INChC4}};
  \code{\link{INRWC1}}; \code{\link{INRWC2}};
  \code{\link{RWC_EV1}}; \code{\link{RWC_EV2}};
  \code{\link{RWC_EV3}}; \code{\link{inplahe1}};
  \code{\link{inplahe2}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

