% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INChC3.R
\name{INChC3}
\alias{INChC3}
\title{Function for calculating the Increase Chlorophyll Content 3 - Third difference (INChC3)}
\usage{
INChC3(ChC4, ChC3)
}
\arguments{
\item{ChC4}{Chlorophyll Content sample 4}

\item{ChC3}{Chlorophyll Content sample 3}
}
\value{
Return the Increase Chlorophyll Content 3
}
\description{
Function for calculating the Increase Chlorophyll Content 3 - Third difference (INChC3)
}
\details{
This function calculates the Increase Chlorophyll Content 3 (Third difference)
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}},
  \code{\link{INChC2}}, \code{\link{INChC4}},
  \code{\link{INRWC1}}, \code{\link{INRWC2}},
  \code{\link{RWC_EV1}}, \code{\link{RWC_EV2}},
  \code{\link{RWC_EV3}}, \code{\link{inplahe1}},
  \code{\link{inplahe2}}, \code{\link{inplahe3}},
  \code{\link{insd1}}, \code{\link{insd2}},
  \code{\link{insnpp1}}, \code{\link{insnpp2}}
}
\author{
Omar Benites
}
