% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INChC4.R
\name{INChC4}
\alias{INChC4}
\title{Function for calculating the Increase Chlorophyll Content 4 - Fourth difference (INChC4)}
\usage{
INChC4(ChC5, ChC4)
}
\arguments{
\item{ChC5}{Chlorophyll Content sample 5}

\item{ChC4}{Chlorophyll Content sample 4}
}
\value{
Return the Increase Chlorophyll Content 4
}
\description{
Function for calculating the Increase Chlorophyll Content 4 - Fourth difference (INChC4)
}
\details{
This function calculates the Increase Chlorophyll Content 4 (Fourth difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}},
  \code{\link{INChC2}}, \code{\link{INChC3}},
  \code{\link{INRWC1}}, \code{\link{INRWC2}},
  \code{\link{RWC_EV1}}, \code{\link{RWC_EV2}},
  \code{\link{RWC_EV3}}, \code{\link{inplahe1}},
  \code{\link{inplahe2}}, \code{\link{inplahe3}},
  \code{\link{insd1}}, \code{\link{insd2}},
  \code{\link{insnpp1}}, \code{\link{insnpp2}}
}

