% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inplahe1.R
\name{inplahe1}
\alias{inplahe1}
\title{Function for calculating the Increase Plant height 1 (INPLAHE1)}
\usage{
inplahe1(plahe_ev1, plahe_ev2)
}
\arguments{
\item{plahe_ev1}{Plant height 1 Evaluation Days After Planting (DAP)}

\item{plahe_ev2}{Plant height 2 Evaluation Days After Planting (DAP)}
}
\value{
inplahe1 Return the increase plant height 1 (First difference)
}
\description{
Function for calculating the Increase Plant height 1 (INPLAHE1)
}
\details{
This function returns the increase plant height 1 (First difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{RWC_EV3}};
  \code{\link{inplahe2}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

