% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inplahe2.R
\name{inplahe2}
\alias{inplahe2}
\title{Function for calculating the Increase Plant height 2 (INPLAHE2)}
\usage{
inplahe2(plahe_ev2, plahe_ev3)
}
\arguments{
\item{plahe_ev2}{Plant height 2 Evaluation Days After Planting (DAP)}

\item{plahe_ev3}{Plant height 3 Evaluation Days After Planting (DAP)}
}
\value{
inplahe2 Return the increase plant height 2 (Second difference)
}
\description{
Function for calculating the Increase Plant height 2 (INPLAHE2)
}
\details{
This function calculate the increase plant height 2 (Second difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{RWC_EV3}};
  \code{\link{inplahe1}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

