% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inplahe3.R
\name{inplahe3}
\alias{inplahe3}
\title{Function for calculating the Increase Plant height 3 (INPLAHE3)}
\usage{
inplahe3(plahe_ev3, plahe_ev4)
}
\arguments{
\item{plahe_ev3}{Plant height 3 Evaluation Days After Planting (DAP)}

\item{plahe_ev4}{Plant height 4 Evaluation Days After Planting (DAP)}
}
\value{
inplahe3 Return the increase plant height 3 (Third difference)
}
\description{
Function for calculating the Increase Plant height 3 (INPLAHE3)
}
\details{
This function calculates the increase plant height 3 (Third difference)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{RWC_EV3}};
  \code{\link{inplahe1}}; \code{\link{inplahe2}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

