% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/INRWC1.R
\name{INRWC1}
\alias{INRWC1}
\title{Function for calculating Increase the Relative Water Content 1 (INRWC1)}
\usage{
INRWC1(RWC_EV1, RWC_EV2)
}
\arguments{
\item{RWC_EV1}{Relative water content 1 evaluation-First Evaluation (RWC_EV1)}

\item{RWC_EV2}{Relative water content 2 evaluation-First Evaluation (RWC_EV2)}
}
\value{
Return the Increase Relative Water Content 1 (INRWC1)
}
\description{
Function for calculating Increase the Relative Water Content 1 (INRWC1)
}
\details{
This function calculates the Increase Relative Water Content 1 (INRWC1)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC2}};
  \code{\link{RWC_EV1}}; \code{\link{RWC_EV2}};
  \code{\link{RWC_EV3}}; \code{\link{inplahe1}};
  \code{\link{inplahe2}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

