% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/INRWC2.R
\name{INRWC2}
\alias{INRWC2}
\title{Function for calculating Increase Relative Water Content 2 (INRWC2)}
\usage{
INRWC2(RWC_EV2, RWC_EV3)
}
\arguments{
\item{RWC_EV2}{Relative water content 2 evaluation-First Evaluation (RWC_EV2)}

\item{RWC_EV3}{Relative water content 3 evaluation-First Evaluation (RWC_EV3)}
}
\value{
Return the Increase Relative Water Content 2 (INRWC2)
}
\description{
Function for calculating Increase Relative Water Content 2 (INRWC2)
}
\details{
This function calculates the Increase Relative Water Content 2 (INRWC2)
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{RWC_EV1}}; \code{\link{RWC_EV2}};
  \code{\link{RWC_EV3}}; \code{\link{inplahe1}};
  \code{\link{inplahe2}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

