% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/insd1.R
\name{insd1}
\alias{insd1}
\title{Function for calculating the Increase Stem diameter 1 - First difference (INSD1)}
\usage{
insd1(sd_ev1, sd_ev2)
}
\arguments{
\item{sd_ev1}{Stem diameter first Evaluation DAP}

\item{sd_ev2}{Stem diameter second Evaluation DAP}
}
\value{
insd1 Return the Increase Stem diameter 1 - First difference
}
\description{
Function for calculating the Increase Stem diameter 1 - First difference (INSD1)
}
\details{
This function calculates the Increase Stem diameter 1 - First difference
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{RWC_EV3}};
  \code{\link{inplahe1}}; \code{\link{inplahe2}};
  \code{\link{inplahe3}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

