% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insd2.R
\name{insd2}
\alias{insd2}
\title{Function for calculating the Increase Stem diameter 2 - Second difference (INSD2)}
\usage{
insd2(sd_ev2, sd_ev3)
}
\arguments{
\item{sd_ev2}{Stem diameter second Evaluation DAP}

\item{sd_ev3}{Stem diameter third Evaluation DAP}
}
\value{
insd2 Return the Increase Stem diameter 2 - Second difference
}
\description{
Function for calculating the Increase Stem diameter 2 - Second difference (INSD2)
}
\details{
This function calculates the Increase Stem diameter 2 - First difference
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}},
  \code{\link{INChC2}}, \code{\link{INChC3}},
  \code{\link{INChC4}}, \code{\link{INRWC1}},
  \code{\link{INRWC2}}, \code{\link{RWC_EV1}},
  \code{\link{RWC_EV2}}, \code{\link{RWC_EV3}},
  \code{\link{inplahe1}}, \code{\link{inplahe2}},
  \code{\link{inplahe3}}, \code{\link{insd1}},
  \code{\link{insnpp1}}, \code{\link{insnpp2}}
}
\author{
Omar Benites
}
