% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/insnpp1.R
\name{insnpp1}
\alias{insnpp1}
\title{Function for calculating the Increase Stem number/plant 1 (INSNPP1)}
\usage{
insnpp1(snpp_ev1, snpp_ev2)
}
\arguments{
\item{snpp_ev1}{Stem number/plant 1 Evaluation DAP}

\item{snpp_ev2}{Stem number/plant 2 Evaluation DAP}
}
\value{
insnpp1 Return the Increase Stem number/plant 1
}
\description{
Function for calculating the Increase Stem number/plant 1 (INSNPP1)
}
\details{
This function calculates the Increase Stem number/plant 1
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{RWC_EV3}};
  \code{\link{inplahe1}}; \code{\link{inplahe2}};
  \code{\link{inplahe3}}; \code{\link{insd1}};
  \code{\link{insd2}}; \code{\link{insnpp2}}
}

