% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_contained}
\alias{is_contained}
\title{Logical function to detect if a group of elements are part of the set of header}
\usage{
is_contained(..., set)
}
\description{
Check if a header(s) correspond to the fieldbook names. It returns a TRUE if all values are contained in the set and FALSE
if just few or neither elements are contained into the set.
}

