% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtwpl.R
\name{mtwpl}
\alias{mtwpl}
\title{Formula for calculating the Marketable Tuber Weight/Plant (MTWPL)}
\usage{
mtwpl(mtwp, nph)
}
\arguments{
\item{mtwp}{Marketable tuber weight/plot}

\item{nph}{Number of plants harvested}
}
\value{
Return the Marketable Tuber Weight/Plot
}
\description{
Formula for calculating the Marketable Tuber Weight/Plant (MTWPL)
}
\details{
Formula for calculating the Marketable tuber weight/plant
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest,plant.: \code{\link{mtwp}},
  \code{\link{ttya}}, \code{\link{ttyna}}
}
\author{
Omar Benites
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}
