% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mtyna.R
\name{mtyna}
\alias{mtyna}
\title{Formula for calculating the Marketable Tuber Yield No Adjusted (MTYNA)}
\usage{
mtyna(mtwp, pls)
}
\arguments{
\item{mtwp}{Marketable tuber weight/plot}

\item{pls}{Net plot size}
}
\value{
Return the Total tuber yield no adjusted
}
\description{
Formula for calculating the Marketable Tuber Yield No Adjusted (MTYNA)
}
\details{
Formula for calculating the total tuber yield no adjusted
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\keyword{agronomy,harvest,quantitative-continuous,plant}
\keyword{potato,}

