% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nmtp.R
\name{nmtp}
\alias{nmtp}
\title{Formula for calculating the Number of Marketable Tubers/Plot (NMTP)}
\usage{
nmtp(nmtci, nmtcii, nmtciii)
}
\arguments{
\item{nmtci}{Number marketable tubers category I/plot}

\item{nmtcii}{Number marketable tubers category II/plot}

\item{nmtciii}{Number marketable tubers category III/plot}
}
\value{
Return the Number of Marketable Tubers/Plot
}
\description{
Formula for calculating the Number of Marketable Tubers/Plot (NMTP)
}
\details{
Formula for calculating the Number Marketable Tubers/Plot NMTP
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\keyword{agronomy,harvest,quantitative-continuous,yield,}
\keyword{blight,plot}
\keyword{late}
\keyword{potato,}

