% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtpl.R
\name{nmtpl}
\alias{nmtpl}
\title{Formula for calculating the Number of Marketable Tubers/Plant (NMTPL)}
\usage{
nmtpl(nmtp, nph = NA)
}
\arguments{
\item{nmtp}{Number of marketable tubers/plot}

\item{nph}{Number of plants harvested}
}
\value{
Return the
}
\description{
Formula for calculating the Number of Marketable Tubers/Plant (NMTPL)
}
\details{
Formula for calculating the number marketable tubers/plant
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest,plant: \code{\link{atmw}},
  \code{\link{mtya}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

