% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrpp.R
\name{nrpp}
\alias{nrpp}
\title{Formula for calculating the Number of Roots per Plant (NRPP)}
\usage{
nrpp(nonc, nocr, noph)
}
\arguments{
\item{nonc}{Number of non commercial roots}

\item{nocr}{Number of commercial roots}

\item{noph}{Number of plants harvested}
}
\value{
Return the number of roots per plant 'nrpp'
}
\description{
Formula for calculating the Number of Roots per Plant (NRPP)
}
\details{
Formula for calculating the number of roots per plant
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}},
  \code{\link{biom}}, \code{\link{ci}},
  \code{\link{cytha}}, \code{\link{dmfy}},
  \code{\link{dmry}}, \code{\link{dm}},
  \code{\link{fytha}}, \code{\link{hi}}, \code{\link{rfr}},
  \code{\link{rytha}}, \code{\link{trw}}, \code{\link{ypp}}
}
\author{
Omar Benites, Raul Eyzaguirre
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}
