% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs1.R
\name{ocs1}
\alias{ocs1}
\title{Formula for calculating the Oil Content Sample 1 Percentage (OCS1)}
\usage{
ocs1(iws1, fws1)
}
\arguments{
\item{iws1}{Initial weight sample 1}

\item{fws1}{Fresh weight of tuber sample 1}
}
\value{
Return the Oil content sample 1 percentage
}
\description{
Formula for calculating the Oil Content Sample 1 Percentage (OCS1)
}
\details{
Formula for calculating the Oil content sample 1 percentage
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,post-harvest,plant: \code{\link{aocp}},
  \code{\link{atw}}, \code{\link{avdm}}, \code{\link{dm1}},
  \code{\link{dm2}}, \code{\link{ocs2}}, \code{\link{sg}}
}
\author{
Omar Benites
}
\keyword{agronomy,post-harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}
