% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outlier_print.R
\name{outlier_colors}
\alias{outlier_colors}
\title{Function to print outlier with colours}
\usage{
outlier_colors(file = "mydata.xlsx", trait, sheetname = "mysheet", ll, ul)
}
\arguments{
\item{file}{filename of the fieldbook}

\item{trait}{Name of the trait column}

\item{sheetname}{Sheet name of the fieldbook}

\item{ll}{lower limit}

\item{ul}{upper limit}
}
\description{
Function to print outlier with colours
}

