% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_if.R
\name{pct_if}
\alias{pct_if}
\title{Percentage of increasing or decreasing of trait values.}
\usage{
pct_if(inital_value = NA, final_value = NA)
}
\arguments{
\item{inital_value}{The initial value of the trait}

\item{final_value}{The final value of the trait}
}
\value{
Return A vector with the diference with the initial and final trait values measured in percentage
}
\description{
At the post-harvest stage in mother and baby trials, it tipical to measure the gain or loss values given by
field traits.
}
\author{
Omar Benites
}
\references{
Participatory Varietal Selection (PVS) Mother and baby trial design. International Potato Center, 2014.
}

