% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pph.R
\name{pph}
\alias{pph}
\title{Formula for calculating the Percentage of Plants Harvested (PPH)}
\usage{
pph(nph, ntp)
}
\arguments{
\item{nph}{Number of plants harvested}

\item{ntp}{Number of tubers planted}
}
\value{
A vector that contains the percentage of plants harvested 'pph'
}
\description{
Formula to calculate the Percentage of Plants Harvested.
}
\details{
Formula to calculate the percentage of plants harvested
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,Yield; Late Blight: \code{\link{shi}}
}
\author{
Omar Benites
}
\keyword{Blight}
\keyword{Late}
\keyword{agronomy,harvest,quantitative-continuous,percentage,Yield;}
\keyword{potato,}
