% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_spt.R
\name{pw_spt}
\alias{pw_spt}
\title{Function for calculating the Percentage weight loss of sprouted tuber (PW_SPT)}
\usage{
pw_spt(itw, ftw)
}
\arguments{
\item{itw}{Initial tuber weight}

\item{ftw}{Final tuber weight}
}
\value{
Return the percentage weight loss of sprouted tuber
}
\description{
Function for calculating the Percentage weight loss of sprouted tuber (PW_SPT)
}
\details{
This function returns the percentage weight loss of sprouted tuber
}
\references{
Procedures for standard evaluation and data management of advanced potato clones: Assessment of Dormancy and Sprouting
Behavior of Elite and Advanced Clones.International Potato Center (CIP).
}
\seealso{
Other Sprout-dormancy, evaluation, potato: \code{\link{avlglatsp}},
  \code{\link{avthsp}}, \code{\link{dorpd}},
  \code{\link{pw_uspt}}
}
\author{
Omar Benites
}
