% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pw_uspt.R
\name{pw_uspt}
\alias{pw_uspt}
\title{Function for calculating the percentage weight loss of unsprouted tuber (PW_USPT)}
\usage{
pw_uspt(itw, intw)
}
\arguments{
\item{itw}{Initial tuber weight}

\item{intw}{Intermediate tuber weight}
}
\value{
Return the percentage weight loss of unsprouted tuber
}
\description{
Function for calculating the percentage weight loss of unsprouted tuber (PW_USPT)
}
\details{
This function calculate the percentage weight loss of unsprouted tuber
}
\author{
Omar Benites
}
\references{
Procedures for standard evaluation and data management of advanced potato clones: Assessment of Dormancy and Sprouting
Behavior of Elite and Advanced Clones.International Potato Center (CIP).
}
\seealso{
Other Sprout-dormancy, evaluation, potato: \code{\link{avlglatsp}};
  \code{\link{avthsp}}; \code{\link{dorpd}};
  \code{\link{pw_spt}}
}

