% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfr.R
\name{rfr}
\alias{rfr}
\title{Formula for calculating the Root Foliage Ratio (RFR)}
\usage{
rfr(crw, ncrw, vw, dmvd, dmvf)
}
\arguments{
\item{crw}{Commercial root weight}

\item{ncrw}{Non commercial root weight}

\item{vw}{Vine weight}

\item{dmvd}{Dry weight of DMVD samples}

\item{dmvf}{Fresh weight vines for dry matter assessment}
}
\value{
Return the root foliage ratio 'rfr'
}
\description{
Formula for calculating the Root Foliage Ratio (RFR)
}
\details{
Formula for calculating the root foliage ratio
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{ci}};
  \code{\link{cytha}}; \code{\link{dmfy}};
  \code{\link{dmry}}; \code{\link{dm}};
  \code{\link{fytha}}; \code{\link{hi}};
  \code{\link{nrpp}}; \code{\link{rytha}};
  \code{\link{trw}}; \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

