% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RWC_EV3.R
\name{RWC_EV3}
\alias{RWC_EV3}
\title{Function for calculating Relative Water Content 3 Evaluation - Third Evaluation (RWC_EV3)}
\usage{
RWC_EV3(Leaflet_FW3, Leaflet_DW3, Leaflet_TW3)
}
\arguments{
\item{Leaflet_FW3}{Leaflet fresh weight - 3 Evaluation}

\item{Leaflet_DW3}{Leaflet dry weight weight 3 Evaluation}

\item{Leaflet_TW3}{Leaflet turgid weight - 3 Evaluation}
}
\value{
RWC_EV3 Return the relative water content 3 evaluation-Third Evaluation
}
\description{
Function for calculating Relative Water Content 3 Evaluation - Third Evaluation (RWC_EV3)
}
\details{
This function returns the relative water content 3 evaluation-Third Evaluation
}
\author{
Omar Benites
}
\references{
Assessing Potato Clone for Drought Tolerance under Field Condition
}
\seealso{
Other drought, evaluation, potato: \code{\link{INChC1}};
  \code{\link{INChC2}}; \code{\link{INChC3}};
  \code{\link{INChC4}}; \code{\link{INRWC1}};
  \code{\link{INRWC2}}; \code{\link{RWC_EV1}};
  \code{\link{RWC_EV2}}; \code{\link{inplahe1}};
  \code{\link{inplahe2}}; \code{\link{inplahe3}};
  \code{\link{insd1}}; \code{\link{insd2}};
  \code{\link{insnpp1}}; \code{\link{insnpp2}}
}

