% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{sb_cipnumberquality}
\alias{sb_cipnumberquality}
\title{This function print cipnumber which are not well-written}
\usage{
sb_cipnumberquality(data, cip_name)
}
\arguments{
\item{data}{The fiedlbook data frame}

\item{cip_name}{The name of the column which contains the cipnumbers}
}
\value{
Return all wrong cip numbers
}
\description{
This function print cipnumber which are not well-written
}
\keyword{cipnumber,quality}

