% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sb_fb_summary_excel.R
\name{sb_fbsum_excel}
\alias{sb_fbsum_excel}
\title{Function to calculate summary statistics for fieldbook data from excel files}
\usage{
sb_fbsum_excel(data = NULL, idx, groupfactors = NULL, na.rm = FALSE,
  datadict = NULL)
}
\arguments{
\item{data}{A data frame}

\item{idx}{The name or position of the measured variable that will be summarized.}

\item{groupfactors}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{datadict}{The data frame of the data dictionary for potato and sweetpotato}
}
\value{
A data frame with the count, mean and standard desviation
}
\description{
Function to calculate summary statistics for fieldbook data from excel files
}
\details{
This function is capable of divide the information in categorial or quantitative data based on a data dictionary for potato and sweepotato.
If it is categorical, returns the count #'and mode. And if it is quantitative, returns the count, media and standart desviation.
}
\author{
Omar Benites
}
\references{
Progress in developing a potato ontology for breeders. Reinhard Simon, Vilma Hualla, E. Salas, Rene Gomez, Raul Cordova and Stef de Haan.
Crop Ontology 2014.
}
\seealso{
Other stats,summary: \code{\link{sbmean}};
  \code{\link{sbsummary}}
}
\keyword{stats,}
\keyword{summary}

