% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbnewqualityvar.R
\name{sb_newqualityvar}
\alias{sb_newqualityvar}
\title{Quality control for new fieldbook traits.}
\usage{
sb_newqualityvar(data, genotipes, trait, lfactor = NULL, ll, ul)
}
\arguments{
\item{data}{The data frame of the fieldbook}

\item{genotipes}{The genotipes column name}

\item{trait}{The trait column name}

\item{lfactor}{The factor column name}

\item{ll}{lower limit of values}

\item{ul}{upper limit of values}
}
\description{
Quality control for new fieldbook traits.
}
\details{
This function return the values which are out of range. It does not requiere a data dictionary.
Make data qualit control for new variables using a lower limit and upper defined by users.
}
\author{
Omar Benites
}

