% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sbcalculate.R
\name{sbcalculate}
\alias{sbcalculate}
\title{Calculation of measured variables on potato and sweetpotato}
\usage{
sbcalculate(fb, plot.size = NA, plant.den = NA)
}
\arguments{
\item{fb}{A data frame with the fieldbook data}

\item{plot.size}{The plot size in square meters}

\item{plant.den}{Plant density}
}
\value{
A data frame with the calculated variables from the fieldbook
}
\description{
Calculation of measured variables on potato and sweetpotato
}
\details{
This function allows to calculated several measured trait in one go.
}
\references{
Progress in developing a potato ontology for breeders. Reinhard Simon, Vilma Hualla, Raul Eyzaguirre, Raul Cordova, Robert O M Mwanga, Genoveva Rossel
Progress in developing a sweetpotato ontology for breeders. Reinhard Simon, Vilma Hualla, E. Salas, Rene Gomez, Raul Cordova, Stef de Haan
}

