% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sbmean.R
\name{sbmean}
\alias{sbmean}
\title{Function to calculate the summary by mean for breeding data}
\usage{
sbmean(data = NULL, idx, groupvars = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{idx}{The name or position of the measured variable that will be summarized.}

\item{groupvars}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}
}
\value{
A data frame with the count, mean and standard desviation
}
\description{
Function to calculate the summary by mean for breeding data
}
\details{
This function returns the summary by mean.
Furthermore, this function use the doBy package.
}
\author{
Omar Benites
}
\references{
Some code for this function were inspired by Cookbook for R.
}
\seealso{
Other stats,summary: \code{\link{sbsummary}}
}
\keyword{stats,}
\keyword{summary}

