% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sboxchart.R
\name{sboxcharts}
\alias{sboxcharts}
\title{Boxplots for plant breeding}
\usage{
sboxcharts(data, trait = NA, genotipes = NULL, factors = NULL,
  full_name = NULL, units = NULL)
}
\arguments{
\item{data}{The fieldbook data}

\item{trait}{The column's name of the trait on fieldbook}

\item{genotipes}{The genotipes of the fieldbook}

\item{factors}{The column's name of the factor on fieldbook}

\item{full_name}{The full name of the trait}

\item{units}{The units of the trait (If it is a measure)}
}
\value{
A boxplot chart and the list of outliers
}
\description{
Boxplots for plant breeding
}
\details{
A box chart implementation for plant breeding which allows to complete information about the trait.
If data has a factor (ex. type of irrigation), the function catch the factor labels and plot them. Futhermore,
It returns a list of outliers.
}
\author{
Omar Benites
}

