% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{scriteria_phase}
\alias{scriteria_phase}
\title{Percentages and scores by Plant Development}
\usage{
scriteria_phase(phase, data)
}
\arguments{
\item{phase}{Phase or Stage of the plant development. There are 3 options: Flowering, Harvesting and Storage}

\item{data}{A data frame with the field data}
}
\description{
Percentages and scores by Plant Development
}

