% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sg_1hd.R
\name{sg_1hd}
\alias{sg_1hd}
\title{Function for calculating Specific Gravity - First Harvest Date (SG_1HD)}
\usage{
sg_1hd(sgs1_1hd, sgs2_1hd)
}
\arguments{
\item{sgs1_1hd}{Specific gravity sample 1 (First harvest date)}

\item{sgs2_1hd}{Specific gravity sample 2 (First harvest date)}
}
\value{
sg_1hd Return the specific gravity (First harvest date)
}
\description{
Function for calculating Specific Gravity - First Harvest Date (SG_1HD)
}
\details{
This function returns the specific gravity (First harvest date)
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}};
  \code{\link{atmw_2hd}}; \code{\link{atmw_3hd}};
  \code{\link{atnomw_1hd}}; \code{\link{atnomw_2hd}};
  \code{\link{atnomw_3hd}}; \code{\link{sg_2hd}};
  \code{\link{sg_3hd}}; \code{\link{sgs1_1hd}};
  \code{\link{sgs1_2hd}}; \code{\link{sgs1_3hd}};
  \code{\link{sgs2_1hd}}; \code{\link{sgs2_2hd}};
  \code{\link{sgs2_3hd}}
}

