% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_2hd.R
\name{sg_2hd}
\alias{sg_2hd}
\title{Function for calculating Specific Gravity - Second Harvest Date (SG_2HD)}
\usage{
sg_2hd(sgs1_2hd, sgs2_2hd)
}
\arguments{
\item{sgs1_2hd}{Specific gravity sample 1 (Second harvest date)}

\item{sgs2_2hd}{Specific gravity sample 2 (Second harvest date)}
}
\value{
sg_2hd Return the specific gravity (Second harvest date)
}
\description{
Function for calculating Specific Gravity - Second Harvest Date (SG_2HD)
}
\details{
This function returns the specific gravity (Second harvest date)
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{atnomw_3hd}}, \code{\link{sg_1hd}},
  \code{\link{sg_3hd}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}

