% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_3hd.R
\name{sg_3hd}
\alias{sg_3hd}
\title{Function for calculating Specific Gravity - Third Harvest Date (SG_2HD)}
\usage{
sg_3hd(sgs1_3hd, sgs2_3hd)
}
\arguments{
\item{sgs1_3hd}{Specific gravity sample 1 (Third harvest date)}

\item{sgs2_3hd}{Specific gravity sample 2 (Third harvest date)}
}
\value{
sg_2hd Return the specific gravity (Third harvest date)
}
\description{
Function for calculating Specific Gravity - Third Harvest Date (SG_2HD)
}
\details{
This function returns the specific gravity (Third harvest date)
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{atnomw_3hd}}, \code{\link{atnomw}},
  \code{\link{av_sg}}, \code{\link{sg_2hd}},
  \code{\link{sg_average}}, \code{\link{sgs1_1hd}},
  \code{\link{sgs1_2hd}}, \code{\link{sgs1_3hd}},
  \code{\link{sgs2_1hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}
\author{
Omar Benites
}
