% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_formulas.R
\name{sglo}
\alias{sglo}
\title{Plot Mother and Baby Score Global (SGLO)}
\usage{
sglo(sm = NULL, sw = NULL)
}
\arguments{
\item{sm}{Plot Mother/Baby Score by Women. (Number of corn beans)}

\item{swm}{Plot Mother/Baby Score by Men. (Number of corn kernels)}
}
\description{
In Participatory Varietal Selection (PVS) is quite common calculate Plot Mother Score global which is the sum of 
Plot Mother/Baby Score by Men (sm) and Plot Mother/Baby Score by Women (sw). All these scores are provided by agricultors using the
number of corn kernels or beans for every genotype in the field.
}
\examples{
 
\dontrun{
#Participatory data example
sdata <- data.frame(genotypes= c("cip1","cip2","cip3","cip4"), score_women= c(11,6,8,4), score_men = c(10,5,10,1))
#The global score : score_women + score_men
sglo  <-  sglo(sdata$score_women,sdata$score_men)
}
 
}
\author{
Omar Benites
}
\references{
Participatory Varietal Selection (PVS): Mother and baby trial design. International Potato Center. 2014
}

