% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs2_1hd.R
\name{sgs2_1hd}
\alias{sgs2_1hd}
\title{Function for calculating  Specific Gravity Sample 2 - First Harvest Date (SGS2_1HD)}
\usage{
sgs2_1hd(twa1_s2, tww1_s2)
}
\arguments{
\item{twa1_s2}{Tuber weight in air sample 2 (First Harvest Date)}

\item{tww1_s2}{Tuber weight in water sample 2 (First Harvest Date)}
}
\value{
sgs2_1hd Return the specific Gravity sample 2 - First Harvest Date
}
\description{
Function for calculating  Specific Gravity Sample 2 - First Harvest Date (SGS2_1HD)
}
\details{
This function returns specific gravity sample 2 - first harvest date
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2014
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}},
  \code{\link{atmw_2hd}}, \code{\link{atmw_3hd}},
  \code{\link{atnomw_1hd}}, \code{\link{atnomw_2hd}},
  \code{\link{atnomw_3hd}}, \code{\link{atnomw}},
  \code{\link{av_sg}}, \code{\link{sg_2hd}},
  \code{\link{sg_3hd}}, \code{\link{sg_average}},
  \code{\link{sgs1_1hd}}, \code{\link{sgs1_2hd}},
  \code{\link{sgs1_3hd}}, \code{\link{sgs2_2hd}},
  \code{\link{sgs2_3hd}}
}
\author{
Omar Benites
}
