% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sgs2_2hd.R
\name{sgs2_2hd}
\alias{sgs2_2hd}
\title{Function for calculating  Specific Gravity Sample 2 - Second Harvest Date (SGS2_2HD)}
\usage{
sgs2_2hd(twa2_s2, tww2_s2)
}
\arguments{
\item{twa2_s2}{Tuber weight in air sample 2 (Second Harvest Date)}

\item{tww2_s2}{Tuber weight in water sample 2 (Second Harvest Date)}
}
\value{
sgs2_2hd Return the specific Gravity sample 2 - Second Harvest Date
}
\description{
Function for calculating  Specific Gravity Sample 2 - Second Harvest Date (SGS2_2HD)
}
\details{
This function returns specific gravity sample 2 - Second harvest date
}
\author{
Omar Benites
}
\references{
Protocol for tuber bulking maturity assessment of elite and advanced potato clones. International Potato Center (CIP), 2024
}
\seealso{
Other Bulking-maturity, evaluation, potato: \code{\link{atmw_1hd}};
  \code{\link{atmw_2hd}}; \code{\link{atmw_3hd}};
  \code{\link{atnomw_1hd}}; \code{\link{atnomw_2hd}};
  \code{\link{atnomw_3hd}}; \code{\link{sg_1hd}};
  \code{\link{sg_2hd}}; \code{\link{sg_3hd}};
  \code{\link{sgs1_1hd}}; \code{\link{sgs1_2hd}};
  \code{\link{sgs1_3hd}}; \code{\link{sgs2_1hd}};
  \code{\link{sgs2_3hd}}
}

