% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shi.R
\name{shi}
\alias{shi}
\title{Formula for calculating the Harvest sowing index - Survival (SHI)}
\usage{
shi(noph, nops)
}
\arguments{
\item{noph}{Number of plants harvested}

\item{nops}{Number of plants planted}
}
\value{
A vector that contains the harvest sowing index - Survival 'shi'
}
\description{
Formula for calculating the Harvest sowing index - Survival (SHI)
}
\details{
Formula for calculating the harvest sowing index - Survival
}
\author{
Omar Benites
}
\seealso{
Other potato,Yield; Late Blight: \code{\link{pph}}
}
\keyword{Blight}
\keyword{Late}
\keyword{agronomy,harvest,quantitative-continuous,percentage,Yield;}
\keyword{potato,}

