% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorpd.R
\name{str2days}
\alias{str2days}
\title{Strings to days}
\usage{
str2days(adate = "")
}
\arguments{
\item{adate}{Data in format yyyy-mm-dd}
}
\value{
Return the Julian date
}
\description{
Strings to days
}
\details{
This function converts a date string (yyyy-mm-dd) to julian date (ddMonthNameyyyy)
}
\note{
str2days("2011-09-07") to 7Sep2011
}
\author{
Reinhard Simon
}
