% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumbymean}
\alias{sumbymean}
\title{Resumen of variables to calculate the mean}
\usage{
sumbymean(datos = NULL, idx, groupvars = NULL, na.rm = FALSE)
}
\arguments{
\item{datos}{the fieldbook data}

\item{idx}{the genotypes}

\item{groupvars}{Extra treatments}

\item{na.rm}{Logical. TRUE to remove NA values. FALSE is by default}
}
\description{
Summarization of variables according to institutional number and factors (treatments)
if just few or neither elements are contained into the set.
}
\author{
Omar Benites
}
