% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrplot.R
\name{tnrplot}
\alias{tnrplot}
\title{Formula for calculating the Total number of root counting per plot (TNRPLOT)}
\usage{
tnrplot(nonc, nocr)
}
\arguments{
\item{nonc}{Number of non commercial roots}

\item{nocr}{Number of commercial roots}
}
\value{
Return the number of roots per plant 'tnrplot'
}
\description{
Formula for calculating the Total number of root counting per plot (TNRPLOT)
}
\details{
Formula for calculating total number of root counting per plot
}
\author{
Omar Benites, Raul Eyzaguirre
}
\keyword{agronomy,}
\keyword{harvest,}
\keyword{quantitative-continuous,}
\keyword{sweetpotato,}
\keyword{yield}
