% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp.R
\name{tntp}
\alias{tntp}
\title{Formula for calculating the Total Number of Tubers/Plot (TNTP)}
\usage{
tntp(nmtp, nnomtp, nmtci, nmtcii)
}
\arguments{
\item{nmtp}{Number marketable tubers/plot}

\item{nnomtp}{Number of non-marketable tubers/plot}

\item{nmtci}{Number marketable tubers comercial I}

\item{nmtcii}{Number marketable tubers comercial II}
}
\value{
A vector that contains the Total number of tubers/plot
}
\description{
Formula for calculating the Total Number of Tubers/Plot (TNTP)
}
\details{
Formula for calculating the Total number of tubers/plot TNTP
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest: \code{\link{tntpl}},
  \code{\link{ttwp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield,}
\keyword{blight}
\keyword{late}
\keyword{potato,}

