% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trw.R
\name{trw}
\alias{trw}
\title{Formula for calculating the Total Root Weight (TRW)}
\usage{
trw(crw, ncrw)
}
\arguments{
\item{crw}{Commercial root weight}

\item{ncrw}{Non commercial root weight}
}
\value{
Return the Total root weight 'trw'
}
\description{
Formula for calculating the Total Root Weight (TRW)
}
\details{
Formula for calculating the total root weight
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}},
  \code{\link{biom}}, \code{\link{ci}},
  \code{\link{cytha}}, \code{\link{dmfy}},
  \code{\link{dmry}}, \code{\link{dm}},
  \code{\link{fytha}}, \code{\link{hi}},
  \code{\link{nrpp}}, \code{\link{rfr}},
  \code{\link{rytha}}, \code{\link{ypp}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield,}
\keyword{sweetpotato,}

