% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttwp.R
\name{ttwp}
\alias{ttwp}
\title{Formula for Total Tuber Weight/Plot (TTWP)}
\usage{
ttwp(mtwci, mtwcii, mtwp, nomtwp)
}
\arguments{
\item{mtwci}{Marketable tuber weight category I/plot}

\item{mtwcii}{Marketable tuber weight category II/plot}

\item{mtwp}{Markeable tuber weight/plot}

\item{nomtwp}{Non-marketable tuber weight/plot}
}
\value{
Return the Total tuber weight/plot 'ttwp'
}
\description{
Formula for Total Tuber Weight/Plot (TTWP)
}
\details{
This functions calculate Total tuber weight/plot
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest: \code{\link{tntpl}},
  \code{\link{tntp}}
}
\author{
Omar Benites
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plot}
\keyword{potato,}
