% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttwpl.R
\name{ttwpl}
\alias{ttwpl}
\title{Formula for calculating the Total tuber weight/plant (TTWPL)}
\usage{
ttwpl(ttwp, nph)
}
\arguments{
\item{ttwp}{Total tuber weight/plot}

\item{nph}{Number of plants harvested}
}
\value{
Return the
}
\description{
Formula for calculating the Total tuber weight/plant (TTWPL)
}
\details{
Formula for calculating the Total tuber weight/plant
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\author{
Omar Benites
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plot}
\keyword{potato,}
