% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttya.R
\name{ttya}
\alias{ttya}
\title{Formula for calculating the Total Tuber Yield Adjusted (TTYA)}
\usage{
ttya(ttwpl, plantden)
}
\arguments{
\item{ttwpl}{Total tuber weight/plant}

\item{plantden}{Plant density}
}
\value{
Return the total tuber yield adjusted
}
\description{
Formula for calculating the Total Tuber Yield Adjusted (TTYA)
}
\details{
Formula for calculating the total tuber yield adjusted
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest,plant.: \code{\link{mtwpl}},
  \code{\link{mtwp}}, \code{\link{ttyna}}
}
\author{
Omar Benites
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}
