% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ttyna.R
\name{ttyna}
\alias{ttyna}
\title{Formula for calculating the Total Tuber Yield No Adjusted (TTYNA)}
\usage{
ttyna(ttwp, pls)
}
\arguments{
\item{ttwp}{Total tuber weight/plot}

\item{pls}{Net plot size}
}
\value{
Return the Total tuber yield no adjusted  ttyna
}
\description{
Formula for calculating the Total Tuber Yield No Adjusted (TTYNA)
}
\details{
Formula for calculating the Total tuber yield no adjusted
}
\author{
Omar Benites
}
\references{
International Potato Center (CIP).2014: Procedures for Standard Evaluation and Data Management of Advanced Potato Clones
Module 2. Healthy Tuber Yield Trials International Cooperators Guide.
}
\seealso{
Other potato,yield,late blight,harvest,plant.: \code{\link{mtwpl}};
  \code{\link{mtwp}}; \code{\link{ttya}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield,late}
\keyword{blight,plant}
\keyword{potato,}

