% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yield.index}
\alias{yield.index}
\title{Yield index}
\usage{
yield.index(treatYield)
}
\arguments{
\item{treatYield}{the yield column}
}
\description{
Summarization of variables according to institutional number and factors (treatments)
if just few or neither elements are contained into the set.
}
\author{
Omar Benites
}
