% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yield.mean.treat}
\alias{yield.mean.treat}
\title{Yield index based on the means}
\usage{
yield.mean.treat(treatYield)
}
\arguments{
\item{treatYield}{the yield column}
}
\description{
Summarization of variables based on the means by treatment
if just few or neither elements are contained into the set.
}
\author{
Omar Benites
}

