% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ypp.R
\name{ypp}
\alias{ypp}
\title{Formula for calculating the Yield Per Plant (YPP)}
\usage{
ypp(crw, ncrw, noph)
}
\arguments{
\item{crw}{Commercial root weight}

\item{ncrw}{Non commercial root weight}

\item{noph}{Number of plants harvested}
}
\value{
Return the yield per plant 'ypp'
}
\description{
Formula for calculating the Yield Per Plant (YPP)
}
\details{
Formula for calculating the yield per plant
}
\author{
Omar Benites
}
\seealso{
Other sweetpotato,quality,harvest: \code{\link{acrw}};
  \code{\link{biom}}; \code{\link{ci}};
  \code{\link{cytha}}; \code{\link{hi}};
  \code{\link{nrpp}}; \code{\link{rytha}};
  \code{\link{trw}}
}
\keyword{agronomy,harvest,quantitative-continuous,yield}
\keyword{sweetpotato,}

