% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avlglatsp.R
\name{avlglatsp}
\alias{avlglatsp}
\title{Function for calculating the Average Length of lateral sprout (AVLGLATSP)}
\usage{
avlglatsp(...)
}
\arguments{
\item{...}{Length of the lateral sprout}
}
\value{
avlgatsp
}
\description{
This functions calculate the average length of lateral sprout
}
\references{
Procedures for standard evaluation and data management of advanced potato clones: Assessment of Dormancy and Sprouting
Behavior of Elite and Advanced Clones.International Potato Center (CIP).
}
\seealso{
Other Sprout-dormancy, evaluation, potato: \code{\link{avthsp}},
  \code{\link{dorpd}}, \code{\link{pw_spt}},
  \code{\link{pw_uspt}}
}
\author{
Omar Benites
}
